+++ packages/elements/src/components/RequestMaker/Request/Editor.tsx
@@ -3,7 +3,7 @@ import cn from 'classnames';
 import { observer } from 'mobx-react-lite';
 import * as React from 'react';
 
-import { useRequestMakerStore } from '../../../hooks/useRequestMaker';
+import { useRequestMakerStore } from '../../../hooks/useRequestMakerStore';
 import { getEnabledParams } from '../../../utils/params';
 import { TabTitle } from '../TabTitle';
 import { RequestBody } from './Body';
