+++ packages/elements/src/components/RequestMaker/Request/Parameters.tsx
@@ -5,7 +5,7 @@ import { get, map } from 'lodash';
 import { observer } from 'mobx-react-lite';
 import * as React from 'react';
 
-import { useRequestMakerStore } from '../../../hooks/useRequestMaker';
+import { useRequestMakerStore } from '../../../hooks/useRequestMakerStore';
 import { IParam, ParamField, ParamType } from '../../../stores/request-maker/types';
 
 type InFocus = {
