+++ packages/elements/src/components/RequestMaker/Response/Headers.tsx
@@ -5,7 +5,7 @@ import { map } from 'lodash';
 import { observer } from 'mobx-react-lite';
 import * as React from 'react';
 
-import { useRequestMakerStore } from '../../../hooks/useRequestMaker';
+import { useRequestMakerStore } from '../../../hooks/useRequestMakerStore';
 import { ViolationsDisplay } from './ViolationsDisplay';
 
 export interface IResponseHeaders {
