+++ packages/elements/src/components/RequestMaker/Response/Viewer.tsx
@@ -9,7 +9,7 @@ import { keys } from 'lodash';
 import { observer } from 'mobx-react-lite';
 import * as React from 'react';
 
-import { useRequestMakerStore } from '../../../hooks/useRequestMaker';
+import { useRequestMakerStore } from '../../../hooks/useRequestMakerStore';
 import { getHttpCodeColor } from '../../../utils/http';
 import { TabTitle } from '../TabTitle';
 import { ResponseBody } from './Body';
