+++ packages/elements/src/components/RequestMaker/__tests__/RequestEditor.spec.tsx
@@ -4,7 +4,7 @@ import { mount, ReactWrapper } from 'enzyme';
 import * as React from 'react';
 import { act } from 'react-dom/test-utils';
 
-import { RequestMakerProvider } from '../../../hooks/useRequestMaker';
+import { RequestMakerProvider } from '../../../hooks/useRequestMakerStore';
 import { RequestMakerStore } from '../../../stores/request-maker';
 import { RequestEditor } from '../Request/Editor';
 import { TabTitle } from '../TabTitle';
