+++ packages/elements/src/components/RequestMaker/__tests__/ResponseBody.spec.tsx
@@ -9,7 +9,7 @@ import { mount, ReactWrapper } from 'enzyme';
 import * as React from 'react';
 import { act } from 'react-dom/test-utils';
 
-import { RequestMakerProvider } from '../../../hooks/useRequestMaker';
+import { RequestMakerProvider } from '../../../hooks/useRequestMakerStore';
 import { RequestMakerStore } from '../../../stores/request-maker';
 import { ResponseStore } from '../../../stores/request-maker/response';
 import { stringToArrayBuffer } from '../../../utils/arrayBuffer';
