+++ packages/elements/src/components/RequestMaker/__tests__/SuggestionBar.spec.tsx
@@ -4,7 +4,7 @@ import { mount, ReactWrapper } from 'enzyme';
 import * as React from 'react';
 import { act } from 'react-dom/test-utils';
 
-import { RequestMakerProvider } from '../../../hooks/useRequestMaker';
+import { RequestMakerProvider } from '../../../hooks/useRequestMakerStore';
 import { SuggestionFunc } from '../../../hooks/useSuggestion';
 import { RequestMakerStore } from '../../../stores/request-maker';
 import { SuggestionBar } from '../SuggestionBar';
